Refactors action handling in PawMenu
Centralizes action creation using a helper function.

This change simplifies the action definitions in `PawMenu` and `TEST_PAW_MENU` by introducing a `createActionButtonModel` function in `pawTools.js`. This promotes code reuse and improves readability by abstracting the creation of action objects.
The state of the mainPad is managed by mainPad component now, so the PawMenu use the menuToggled signal
to communicate when a action that need to close the menu is triggerd.