Refactors map loading and tile creation
Streamlines the map loading process by emitting a single signal for each item snapable tile.
This simplifies the signal handling in the editor and allows for more generic tile creation logic.

The ItemSnapable now handles its own serialization to JSON and can print it.

A signal is emitted for each ItemSnapable, which allows to handle tiles more generically.